requirements.txt
streamlit>=1.31.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
openpyxl>=3.1.2
pillow>=10.0.0
pytesseract>=0.3.10
langchain>=0.1.0
langchain-community>=0.0.20
requests>=2.31.0
tqdm>=4.66.0
pydantic>=2.0.0
python-magic-bin>=0.4.14; platform_system == "Windows"

